+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1236](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1236)
 - Added New Search Page Filters
   [GEOD-1234](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1234)
+- Count the Datapoints before allowing downloads to ensure there are not too many
+  [GEOD-915](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-915)
 
 ### Changed
 - The Search Page Filter Icons reflect the dropdown selection